requirements.txt
flask>=3.0
openai>=1.30
manim>=0.18
python-dotenv>=1.0
psycopg2-binary>=2.9
numpy>=1.26
